udpxy: Optimize multicast address internal passing. Patch by AlexeyS @ wl500g.info
014-addmreq.patch commits:
 2517746 udpxy: Optimize multicast address internal passing. Patch by AlexeyS @ wl500g.info
 7bf0341 udpxy 1.0.23
 ff31dd4 udpxy 1.0.23-7

+++ gateway/udpxy/util.c
@@ -701,7 +701,7 @@ a2time( const char* str, time_t* t, time_t from )
             continue;
         }
 
-        if( !isdigit( str[n] ) ) return ERR_NONDIGIT;
+        if( !isdigit( (int)str[n] ) ) return ERR_NONDIGIT;
 
         if( new_field ) {
             new_field = 0;
@@ -797,7 +797,7 @@ a2double( const char* str, double* pval )
     assert( str );
 
     /* skip to the first */
-    for( p = str; *p && !isalpha(*p); ++p );
+    for( p = str; *p && !isalpha((int)*p); ++p );
 
     if( '\0' != *p ) {
         /* there is a modifier, calculate multiplication
